refactor: isolate custom health identity and trim redundant alias tests
