+++ agent/auxiliary_health.py
@@ -1,0 +1,36 @@
+"""Endpoint identity for auxiliary custom-provider health checks."""
+import contextlib
+from typing import Any, Optional
+
+from hermes_cli.route_identity import normalize_route_base_url
+
+def _unhealthy_cache_key(provider: str, base_url: Optional[str] = None) -> Any:
+    """Provider-wide key, or endpoint-specific key for an explicit custom endpoint."""
+    from agent.auxiliary_client import _normalize_chain_label
+    label = _normalize_chain_label(provider)
+    endpoint = normalize_route_base_url(_custom_health_base_url(provider, base_url))
+    if endpoint:
+        return "custom-endpoint", endpoint
+    return label
+
+
+def _custom_health_base_url(provider: str, explicit_base_url: Optional[str] = None) -> str:
+    """Return the concrete custom endpoint used to scope health and failed-route checks."""
+    from agent.auxiliary_client import _current_custom_base_url
+    explicit = str(explicit_base_url or "").strip()
+    from agent.auxiliary_client import _normalize_chain_label
+    label = _normalize_chain_label(provider)
+    if label == "local/custom":
+        return explicit or _current_custom_base_url()
+    if label.startswith("custom:") and explicit:
+        return explicit
+    with contextlib.suppress(ImportError):
+        from hermes_cli.runtime_provider import _get_named_custom_provider, _resolves_to_custom
+        if _resolves_to_custom(label):
+            return explicit or _current_custom_base_url()
+        entry = _get_named_custom_provider(provider)
+        if entry:
+            return explicit or str(entry.get("base_url") or "").strip()
+    return ""
+
+
